+++ packages/admin/src/form/index.ts
@@ -11,4 +11,4 @@ export * from "./Switch";
 export * from "./TextField";
 export * from "./Select";
 export * from "./FinalFormSelect";
-export * from "./FinalFormRangeSlider";
+export * from "./FinalFormRangeInput";