+++ packages/admin/src/themeAugmentation.ts
@@ -1,7 +1,7 @@
 import { ClearInputButtonThemeProps, CometAdminClearInputButtonClassKeys } from "./common";
 import { CometAdminErrorBoundaryClassKeys, ErrorBoundaryThemeProps } from "./error";
 import {
-    CometAdminFinalFormRangeSliderClassKeys,
+    CometAdminFinalFormRangeInputClassKeys,
     CometAdminFormFieldContainerClassKeys,
     CometAdminFormPaperKeys,
     CometAdminFormSectionKeys,
@@ -35,7 +35,7 @@ declare module "@material-ui/core/styles/overrides" {
         CometAdminFormSection: CometAdminFormSectionKeys;
         CometAdminFixedLeftRightLayout: CometAdminFixedLeftRightLayoutKeys;
         CometAdminBreadcrumbs: CometAdminBreadcrumbsClassKeys;
-        CometAdminFinalFormRangeSlider: CometAdminFinalFormRangeSliderClassKeys;
+        CometAdminFinalFormRangeInput: CometAdminFinalFormRangeInputClassKeys;
     }
 }